model folder moved to /src, package.json scripts added
