+++ package.json
@@ -2,11 +2,14 @@
   "name": "paolo-fox-bot",
   "version": "1.1.0",
   "description": "Questo bot invia un file audio con l'oroscopo del giorno di Paolo Fox",
-  "main": "src/index.js",
+  "main": "dist/index.js",
   "scripts": {
-    "start": "node src/index.js",
-    "dev": "nodemon --inspect src/index.js --exec babel-node",
-    "build": "babel src -d dist",
+    "postinstall": "babel src model -d dist",
+    "clean": "rm -rf dist && mkdir dist",
+    "serve": "node dist/index.js",
+    "start": "nodemon --inspect src/index.js --exec babel-node",
+    "build": "npm run clean && npm run build-server",
+    "build-server": "babel src -d dist",
     "test": "mocha --compilers js:babel-register"
   },
   "dependencies": {
